TechWebLinks: relocated to modules folder and updates related references to module
