+++ src/components/Project.jsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 
 import { IconSmall } from "./common/IconSmall";
 import { WebLink } from "./common/WebLink";
-import { TechWebLinks } from "./TechWebLinks";
+import { TechWebLinks } from "../modules/TechWebLinks";
 
 export const Project = ({ project, externalLinks }) => {
   // Show/hide Project object details functionality
