+++ src/modules/TechWebLinks.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { WebLink } from "./common/WebLink";
+import { WebLink } from "../components/common/WebLink";
 
 let keyArray, i, link;
